+++ stac_check/api_lint.py
@@ -41,6 +41,7 @@ class ApiLinter:
         headers: Optional[Dict] = None,
         verbose: bool = False,
         fast: bool = False,
+        fast_linting: bool = False,
     ):
         self.source = source
         self.object_list_key = object_list_key
@@ -48,6 +49,7 @@ class ApiLinter:
         self.headers = headers or {}
         self.verbose = verbose
         self.fast = fast
+        self.fast_linting = fast_linting
         self.version = None
         self.validator_version = self._get_validator_version()
         self.start_time = time.time()
@@ -164,7 +166,7 @@ class ApiLinter:
         ):
             try:
                 results, total_time, schemas = validate_collection_fast(
-                    self.source, Linter, self.verbose
+                    self.source, Linter, self.verbose, self.fast_linting
                 )
 
                 # Store results and metadata